搜索结果优化

+++ src/main/java/org/xinyo/dao/PoetryDao.java
@@ -25,7 +25,7 @@ public interface PoetryDao {
 
     List<Poetry> findPoetrySpByKeyword(Map<String, Object> params);
 
-    long countTotalPoetryByKeyword(Map<String, Object> params);
+    int countTotalPoetryByKeyword(Map<String, Object> params);
 
     List<Integer> findTop100IdByKeyword(String keyword);
