+++ src/main/java/org/xinyo/service/PoetryService.java
@@ -23,7 +23,7 @@ public interface PoetryService {
 
     List<Poetry> findPoetrySpByKeyword(Map<String, Object> params);
 
-    long countTotalPoetryByKeyword(Map<String, Object> params);
+    int countTotalPoetryByKeyword(Map<String, Object> params);
 
     List<Integer> findTop100IdByKeyword(String keyword);
